Use "py -3" instead of "python"
The `python` command is the old way of running Python on Windows.
It only runs a single version (and it could be any version), and
has to be added to PATH explicitly, either at setup or manually.

The `py` launcher on Windows is always enabled by default (meaning a user
would have to explicitly disable it at install to not get it), and it
can launch multiple different versions, such as `py -2`, `py -3`, `py -2.7`,
and `py -3.6`.

https://docs.python.org/3/using/windows.html#launcher